aimodel/requirements.txt
# --- Web/API stack
fastapi==0.116.1
uvicorn==0.30.6
starlette==0.47.3
pydantic==2.11.9
httpx==0.27.0
anyio==4.11.0
typing_extensions==4.15.0
typing-inspection==0.4.1
python-dotenv==1.1.1
PyYAML==6.0.2
h11==0.16.0
httptools==0.6.4
websockets==15.0.1
watchfiles==1.1.0
sniffio==1.3.1
httpcore==1.0.9
diskcache==5.6.3
# --- LLM runtime
gguf==0.17.1
psutil==7.0.0
nvidia-ml-py3==7.352.0     # harmless on CPU; used for GPU stats if present

# --- RAG core
faiss-cpu==1.8.0.post1
numpy==1.26.4
scipy==1.16.1
scikit-learn==1.7.1
sentence-transformers==3.0.1
transformers==4.55.4
tokenizers==0.21.4

# --- Parsing / docs / OCR / data
beautifulsoup4==4.12.3
lxml==6.0.1
readability-lxml==0.8.1
selectolax==0.3.21
pdfminer.six==20240706
PyPDF2==3.0.1
pypdfium2==4.30.0
python-docx==1.1.2
openpyxl==3.1.5
python-pptx==0.6.23
striprtf==0.0.26
pillow==10.4.0
pytesseract==0.3.13
pandas==2.2.2
tzlocal==5.2
tzdata==2025.2

# --- Search / HTTP helpers
ddgs==9.5.4
requests==2.32.5
urllib3==2.5.0
certifi==2025.8.3
idna==3.10
charset-normalizer==3.4.3
regex==2025.7.34

# --- Auth / crypto / licensing helpers
cryptography==45.0.6
python-jose[cryptography]==3.5.0
PyJWT==2.10.1
passlib[bcrypt]==1.7.4
PyNaCl==1.5.0
keyring==25.6.0
